requirements.txt
langchain~=0.3.14
langchain_community
langchain_openai
langgraph~=0.2.61

#YOLO
torch~=2.5.1
opencv-python~=4.10.0.84
ultralytics~=8.3.58

#Phoenix tracing
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-langchain

python-dotenv~=1.0.1
openai~=1.72.0
grandalf~=0.8

torchvision~=0.20.1
numpy~=1.26.4
pillow~=11.1.0
pydantic~=2.10.4

aiofiles~=24.1.0
langchain-core~=0.3.51
typing_extensions~=4.12.2
langchain-openai~=0.3.12
requests~=2.32.3
pytz~=2024.2

transformers >= 4.45.0
accelerate~=1.6.0
azure-ai-inference~=1.0.0b9
azure-core~=1.33.0
PyYAML~=6.0.2

python-docx~=1.1.2
